HEEDLS-462 Move evaluation summary methods into separate data/service
